Refactorización de los paquetes usando la arquitectura de capas y creación de las interfaces y clases de implementación de las interfaces.
