+++ src/main/java/com/jlanzasg/novabank/repository/ClienteRepositorio.java
@@ -1,0 +1,4 @@
+package com.jlanzasg.novabank.repository;
+
+public interface ClienteRepositorio {
+}
