+++ src/main/java/com/jlanzasg/novabank/repository/CuentaBancariaRepositorio.java
@@ -1,0 +1,5 @@
+package com.jlanzasg.novabank.repository;
+
+public interface CuentaBancariaRepositorio {
+
+}
