+++ src/main/java/com/jlanzasg/novabank/repository/CuentaBancariaRepositorioJdbc.java
@@ -1,0 +1,4 @@
+package com.jlanzasg.novabank.repository;
+
+public class CuentaBancariaRepositorioJdbc {
+}
