+++ src/main/java/com/jlanzasg/novabank/service/ClienteService.java
@@ -1,0 +1,4 @@
+package com.jlanzasg.novabank.service;
+
+public interface ClienteService {
+}
