+++ src/test/java/com/jlanzasg/novabank/service/BancoTest.java
@@ -1,7 +1,7 @@
-package com.jlanzasg.novabank.negocio;
+package com.jlanzasg.novabank.service;
 
-import com.jlanzasg.novabank.modelo.Cliente;
-import com.jlanzasg.novabank.modelo.CuentaBancaria;
+import com.jlanzasg.novabank.model.Cliente;
+import com.jlanzasg.novabank.model.CuentaBancaria;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
