+++ src/test/java/com/jlanzasg/novabank/view/MenuOperacionesTest.java
@@ -1,8 +1,8 @@
-package com.jlanzasg.novabank.vista;
+package com.jlanzasg.novabank.view;
 
-import com.jlanzasg.novabank.modelo.Cliente;
-import com.jlanzasg.novabank.modelo.CuentaBancaria;
-import com.jlanzasg.novabank.negocio.Banco;
+import com.jlanzasg.novabank.model.Cliente;
+import com.jlanzasg.novabank.model.CuentaBancaria;
+import com.jlanzasg.novabank.service.Banco;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
